requirements.txt
Flask==2.1.2
numpy==1.23.5
python-dotenv==0.19.2
openai==1.30.5
werkzeug==2.1.2
moviepy==1.0.3
bs4==0.0.2
opencv-python==4.9.0.80

